Push the NLU (en) Bluemix doc to production
- Allen made changes to the Bluemix getting started doc for NLU. I'm
pushing them to prod from GHE Bluemix-Docs repo to GH IBM-Bluemix-Docs
repo. The changes:
-- Removed index.md, which is replaced with getting-started.md
-- Replaced the toc from GH with the toc from GHE that contains Allen's
changes
-- Added getting-started.md which contains the updated text from Allen.
The text is a reworking of the NLU tutorial that was published in the
WDC doc for NLU.

- Note that the PR 1 from customer joeduffi for the syntax errors in the
curl command that pertain to json files is now moot because that text
doesn't exist in the getting-started.md file. I confirmed with Allen and
Garrett.